Change the way the Credential,Validate,Proxy classes handle their received blocks.
This change has been made because the use of instance_eval resulted in the blocks passed to the action methods like success/failure being executed in the scope of the class rather than the Sinatra application. An alternative solution might be possible by defining methods on the @app passed to the class but at this time the quickest solution is to have the class initializer #call the block passing itself as the first parameter.

For details on this issue and suitable tests for playing with solutions please see http://gist.github.com/449614