Change the jQuery URL from http://ajax.googleapis.com/ajax/libs/jquery/1.4.2/jquery.min.js to a local copy.
The reason for using the ajax.googleapis.com URL in the first place is to take advantage of browser caching to speed up the response time. Unfortunately only Safari and Chrome have a positive view of this mixed HTTPS/HTTP content. Other browsers such as Firefox and MSIE raise a warning to the user asking if they want to show the page as it has mixed content. Due to the behaviour of these browsers we've opted for a local copy of the javascript file which will be served over HTTPS and therefor have no caching at all.